refactor: extract prolly_btree_cursor_count.c
Move cursor count helpers and btree count APIs into a dedicated
translation unit. Promote sortKeyFromUnpackedForCount for the seek
path still in prolly_btree_cursor.c. Wire the new file into PROLLY_OBJS,
SRC, the amalgamation engine block, and lint presence/size checks.
